Applied remarks from issue-16,17 PR; Updated javadoc
